Second Lane and Up/Down objects
Added a second lane above the original lane and created two sample objects, one takes up arrow and the other takes down arrow to deactivate. I only created two sample objects so that they can be copied and organized into a pattern that best fits the song. Everything else about the scene should be identical to the original Kira scene as of the time I cloned it.